Fixed the task "prod" to work only if --prod is defined
My goal was to use gulpIf to conditional run a set of tasks
without having to specify gulpIf on each task. This does not look
possible.

While I can use "through2" to wrap a function around a vinyl stream,
I was not able to link multiple tasks such as Minimize, Concat, ...
via the pipe operator